requirements.txt
asgiref==3.9.0
Babel==2.13.1
Django==4.2.5
django-cors-headers==4.3.1
django-extra-views==0.14.0
django-filter==23.3
django-graphql-jwt==0.4.0
django-haystack==3.2.1
django-oscar==4.0
django-phonenumber-field==7.2.0
django-tables2==2.7.0
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
factory_boy==3.3.3
Faker==37.4.0
graphene==3.4.3
graphene-django==3.0.2
graphql-core==3.2.6
graphql-relay==3.2.0
packaging==25.0
phonenumbers==8.13.55
Pillow==10.0.1
promise==2.3
psycopg2-binary==2.9.10
purl==1.6
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
setuptools==80.9.0
six==1.17.0
sorl-thumbnail==12.10.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.14.0
tzdata==2025.2
